requirements.txt
streamlit
pypdf2
langchain
python_dotenv
openai
huggingface_hub
faiss-cpu
